LPS-51438 Fix tests

+++ portal-impl/test/unit/com/liferay/portal/expression/ExpressionEvaluationTest.java
@@ -63,7 +63,8 @@ public class ExpressionEvaluationTest {
 
 		double var3 = var1 + var2;
 
-		Assert.assertEquals(var1 + var2 + var3, expression.evaluate());
+		Assert.assertEquals(
+			(Double)(var1 + var2 + var3), expression.evaluate());
 	}
 
 	@Test
@@ -96,7 +97,7 @@ public class ExpressionEvaluationTest {
 
 		float var3 = var1 + var2;
 
-		Assert.assertEquals(var1 + var2 + var3, expression.evaluate());
+		Assert.assertEquals((Float)(var1 + var2 + var3), expression.evaluate());
 	}
 
 	@Test
@@ -208,7 +209,7 @@ public class ExpressionEvaluationTest {
 		double var3 = var1 + var2;
 
 		Assert.assertEquals(
-			MathUtil.sum(var1, var2, var3), expression.evaluate());
+			(Double)MathUtil.sum(var1, var2, var3), expression.evaluate());
 	}
 
 	@Test
